chore(clerk-js): Move Legal consent checkbox to its own component
